requirements.txt
fastapi
uvicorn
tensorflow
Pillow
python-multipart
SQLAlchemy
bcrypt
pydantic[email]
gdown
requests
python-dotenv
